Step the master FrameSequencer, and the subsequent VolumeEnvelope.
The timing of this isn't exact, since we might actually want to modify the volume, etc. during a single call to APU::step.. Since this is just done for a single instruction at a time, this should be sufficient even if our resulting waveform might be off by one waveform or so.